PolygonDemo/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
polygon-api-client==0.2.11
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.3
